+++ fabric/src/main/java/koala/fishingreal/FishingRealFabric.java
@@ -1,5 +1,6 @@
 package koala.fishingreal;
 
+import fuzs.forgeconfigapiport.fabric.api.neoforge.v4.NeoForgeConfigRegistry;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
@@ -7,6 +8,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.util.profiling.ProfilerFiller;
+import net.neoforged.fml.config.ModConfig;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
@@ -25,5 +27,7 @@ public class FishingRealFabric implements ModInitializer {
                 return listener.reload(synchronizer, manager, prepareProfiler, applyProfiler, prepareExecutor, applyExecutor);
             }
         }));
+        
+        NeoForgeConfigRegistry.INSTANCE.register(FishingReal.MOD_ID, ModConfig.Type.COMMON, Config.CONFIG_SPEC);
     }
 }